Fix semantic native build closure and invariant count

+++ generic_chess/_native/native_semantic_rules.h
@@ -91,7 +91,7 @@ typedef struct {
 typedef struct {
     uint8_t kind;            /* invariant enum code */
     GCSemSquareRef *refs;
-    uint8_t refs_count;
+    uint16_t refs_count;
 } GCSemInvariant;
 
 typedef struct {
